👌 IMPROVE: Secure deletePost route

+++ src/Controller/PostController.php
@@ -71,6 +71,10 @@ class PostController extends AbstractFOSRestController
      */
     public function deletePost(Post $post)
     {
+        if($post->getAuthor() != $this->security->getUser())
+        {
+            return new JsonResponse(['erreur' => 'Vous n\'êtes pas autorisé a faire cette action'], Response::HTTP_BAD_REQUEST);
+        }
         $em = $this->doctrine->getManager();
         $em->remove($post);
         $em->flush();
